add email._policybase
related to https://github.com/python/typeshed/issues/3968
